Se realiza un refactor

+++ Server/includes/Model/DTOs/PlayerDTO.h
@@ -9,9 +9,9 @@ struct PlayerDTO {
     InstanceId player_id;
     int gold;
     int energy;
-    int status;
+    uint8_t status;
 
-    PlayerDTO(InstanceId player_id1, int gold1, int energy1, int status1) : player_id(player_id1), gold(gold1), energy(energy1), status(status1) {}
+    PlayerDTO(InstanceId player_id1, int gold1, int energy1, uint8_t status1) : player_id(player_id1), gold(gold1), energy(energy1), status(status1) {}
 };
